+++ extensions/tidyread---streamline-your-daily-reading/src/utils/preference.ts
@@ -10,22 +10,48 @@ function isNum(value?: string) {
   return isNumber(+value!);
 }
 
+// In order to provide better experience, we should set different default values for different providers. Benchmark is 30 articles.
+const PROVIDER_CONFIG = {
+  openai: {
+    maxApiConcurrency: 10,
+    retryCount: 3,
+    retryDelay: 5,
+  },
+  raycast: {
+    maxApiConcurrency: 5,
+    retryCount: 4,
+    retryDelay: 20,
+  },
+  moonshot: {
+    maxApiConcurrency: 3,
+    retryCount: 3,
+    retryDelay: 5,
+  },
+};
+
+function formatApiModel(pref: Preferences) {
+  return pref.provider === "raycast" ? (pref.apiModel ?? "").replace("raycast-", "") : pref?.apiModel;
+}
+
 export function normalizePreference(): Required<Preferences> {
   const values = getPreferenceValues();
+
+  const defaultApiConfig = PROVIDER_CONFIG[values.provider as keyof typeof PROVIDER_CONFIG];
 
   return {
     provider: values.provider ?? "openai",
     apiKey: values.apiKey || "",
-    apiModel: values.apiModel || "gpt-3.5-turbo-16k",
+    apiModel: formatApiModel(values) || "",
+    maxTokens: isNum(values.maxTokens) ? +values.maxTokens : 200,
     apiHost: values.apiHost || "",
     preferredLanguage: values.preferredLanguage || "",
     httpProxy: values.httpProxy || "",
     summarizePrompt: (values.summarizePrompt || "").trim() || DEFAULT_PROMPT,
     maxItemsPerFeed: isNum(values.maxItemsPerFeed) ? +values.maxItemsPerFeed : 10,
-    maxApiConcurrency: isNum(values.maxApiConcurrency) ? +values.maxApiConcurrency : 3,
-    retryCount: isNum(values.retryCount) ? +values.retryCount : 5,
+    maxApiConcurrency: isNum(values.maxApiConcurrency) ? +values.maxApiConcurrency : defaultApiConfig.maxApiConcurrency,
+    retryCount: isNum(values.retryCount) ? +values.retryCount : defaultApiConfig.retryCount,
     // 默认重试延迟时间为30秒
-    retryDelay: isNum(values.retryDelay) ? +values.retryDelay * 1000 : 30 * 1000,
+    retryDelay: isNum(values.retryDelay) ? +values.retryDelay * 1000 : defaultApiConfig.retryDelay * 1000,
     notificationTime: values.notificationTime || "9am",
     autoGenDigest: values.autoGenDigest ?? false,
     requestTimeout: isNum(values.requestTimeout) ? +values.requestTimeout * 1000 : 20 * 1000,